+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/optimizer/TaxiOptimizationValidation.java
@@ -23,8 +23,8 @@ import java.util.stream.Stream;
 
 import org.matsim.contrib.dvrp.data.Request;
 import org.matsim.contrib.dvrp.fleet.Fleet;
-import org.matsim.contrib.taxi.data.TaxiRequest.TaxiRequestStatus;
-import org.matsim.contrib.taxi.data.TaxiRequests;
+import org.matsim.contrib.taxi.passenger.TaxiRequest.TaxiRequestStatus;
+import org.matsim.contrib.taxi.passenger.TaxiRequests;
 import org.matsim.contrib.taxi.scheduler.TaxiScheduler;
 
 public class TaxiOptimizationValidation {
