+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/optimizer/assignment/AssignmentRequestData.java
@@ -19,7 +19,7 @@
 
 package org.matsim.contrib.taxi.optimizer.assignment;
 
-import org.matsim.contrib.taxi.data.TaxiRequest;
+import org.matsim.contrib.taxi.passenger.TaxiRequest;
 
 import com.google.common.collect.ImmutableList;
