+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/optimizer/assignment/TaxiToRequestAssignmentCostProvider.java
@@ -20,7 +20,7 @@
 package org.matsim.contrib.taxi.optimizer.assignment;
 
 import org.matsim.contrib.dvrp.path.OneToManyPathSearch.PathData;
-import org.matsim.contrib.taxi.data.TaxiRequest;
+import org.matsim.contrib.taxi.passenger.TaxiRequest;
 import org.matsim.contrib.taxi.optimizer.VehicleData;
 import org.matsim.contrib.taxi.optimizer.VehicleData.Entry;
 import org.matsim.contrib.taxi.optimizer.assignment.AssignmentDestinationData.DestEntry;
