+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/optimizer/fifo/FifoSchedulingProblem.java
@@ -23,7 +23,7 @@ import java.util.Collection;
 import java.util.Iterator;
 
 import org.matsim.contrib.dvrp.fleet.Fleet;
-import org.matsim.contrib.taxi.data.TaxiRequest;
+import org.matsim.contrib.taxi.passenger.TaxiRequest;
 import org.matsim.contrib.taxi.optimizer.BestDispatchFinder;
 import org.matsim.contrib.taxi.scheduler.TaxiScheduler;
