+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/passenger/SubmittedTaxiRequestsCollector.java
@@ -25,7 +25,6 @@ import java.util.Map;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.contrib.dvrp.data.Request;
-import org.matsim.contrib.taxi.data.TaxiRequest;
 
 //TODO remove this class once taxi stats are refactored
 public class SubmittedTaxiRequestsCollector {
