+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/schedule/TaxiTaskWithRequest.java
@@ -19,7 +19,7 @@
 
 package org.matsim.contrib.taxi.schedule;
 
-import org.matsim.contrib.taxi.data.TaxiRequest;
+import org.matsim.contrib.taxi.passenger.TaxiRequest;
 
 public interface TaxiTaskWithRequest extends TaxiTask {
 	TaxiRequest getRequest();
