+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/schedule/reconstruct/ScheduleReconstructor.java
@@ -29,7 +29,7 @@ import org.matsim.api.core.v01.population.Person;
 import org.matsim.contrib.dvrp.data.Request;
 import org.matsim.contrib.dvrp.fleet.Fleet;
 import org.matsim.contrib.dvrp.fleet.FleetImpl;
-import org.matsim.contrib.taxi.data.TaxiRequest;
+import org.matsim.contrib.taxi.passenger.TaxiRequest;
 import org.matsim.core.api.experimental.events.EventsManager;
 import org.matsim.core.events.EventsUtils;
 import org.matsim.core.events.MatsimEventsReader;
