+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/util/stats/TaxiStatusTimeProfileCollectorProvider.java
@@ -24,7 +24,7 @@ import java.awt.Paint;
 
 import org.jfree.data.xy.DefaultTableXYDataset;
 import org.matsim.contrib.dvrp.fleet.Fleet;
-import org.matsim.contrib.taxi.data.TaxiRequest.TaxiRequestStatus;
+import org.matsim.contrib.taxi.passenger.TaxiRequest.TaxiRequestStatus;
 import org.matsim.contrib.taxi.passenger.SubmittedTaxiRequestsCollector;
 import org.matsim.contrib.taxi.run.TaxiConfigGroup;
 import org.matsim.contrib.util.timeprofile.TimeProfileCharts;
